requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
pillow
gunicorn
huggingface_hub